+++ src/hvl/projectparmorel/knowledge/Action.java
@@ -13,8 +13,8 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
 
-import hvl.projectparmorel.ml.Error;
-import hvl.projectparmorel.ml.SerializableMethod;
+import hvl.projectparmorel.moderrepair.Error;
+import hvl.projectparmorel.moderrepair.SerializableMethod;
 
 /**
  *
