+++ src/hvl/projectparmorel/moderrepair/ModelFixer.java
@@ -1,4 +1,4 @@
-package hvl.projectparmorel.ml;
+package hvl.projectparmorel.moderrepair;
 
 import java.util.List;
