+++ src/hvl/projectparmorel/moderrepair/ModelProcesser.java
@@ -1,4 +1,4 @@
-  package hvl.projectparmorel.ml;
+  package hvl.projectparmorel.moderrepair;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
