+++ src/hvl/projectparmorel/moderrepair/Preferences.java
@@ -1,4 +1,4 @@
-package hvl.projectparmorel.ml;
+package hvl.projectparmorel.moderrepair;
 
 import java.io.File;
 import java.io.IOException;
