+++ src/hvl/projectparmorel/moderrepair/QModelFixer.java
@@ -1,4 +1,4 @@
-package hvl.projectparmorel.ml;
+package hvl.projectparmorel.moderrepair;
 
 import java.io.IOException;
 import java.util.ArrayList;
