+++ src/hvl/projectparmorel/moderrepair/Sequence.java
@@ -1,4 +1,4 @@
-package hvl.projectparmorel.ml;
+package hvl.projectparmorel.moderrepair;
 
 import java.util.ArrayList;
 import java.util.List;
